+++ cds/variable_length_vector/dense_pointers.cpp
@@ -6,22 +6,22 @@ namespace cds {
     using namespace std;
 
     dense_pointers::dense_pointers(
-        unsigned int max_length, unsigned int rate
+        uint64_t max_length, uint64_t rate
     ) : offsets(ceil(log2((rate - 1) * max_length))) {
         this->rate = rate;
         this->size = 0;
     }
 
-    unsigned int dense_pointers::vector_size() {
+    uint64_t dense_pointers::vector_size() {
         return this->bv.vector_size() +
                this->offsets.vector_size() +
                this->samples.size();
     }
 
-    unsigned int dense_pointers::read(unsigned int index) {
-        unsigned int begin = this->samples[index / this->rate] +
+    uint64_t dense_pointers::read(uint64_t index) {
+        uint64_t begin = this->samples[index / this->rate] +
                              this->offsets.read(index);
-        unsigned int end = this->bv.size;
+        uint64_t end = this->bv.size;
         index++;
 
         if (index < this->offsets.size) {
@@ -33,11 +33,11 @@ namespace cds {
         }
 
         return (
-            this->bv.bits_read(begin, end) | (1 << (end - begin))
+            this->bv.bits_read(begin, end) | (1LL << (end - begin))
         ) - 1;
     }
 
-    void dense_pointers::push_back(unsigned int value) {
+    void dense_pointers::push_back(uint64_t value) {
         if (this->size % this->rate == 0) {
             this->samples.push_back(this->bv.size);
         }
@@ -50,14 +50,14 @@ namespace cds {
         this->size++;
 
         value++;
-        unsigned int value_length = bits_length(value) - 1;
+        uint64_t value_length = bits_length(value) - 1;
         if (value_length == 0) {
             return;
         }
-        value &= ~(1 << value_length);
+        value &= ~(1LL << value_length);
 
-        unsigned int begin = this->bv.size;
-        unsigned int end = begin + value_length;
+        uint64_t begin = this->bv.size;
+        uint64_t end = begin + value_length;
 
         this->bv.resize(end);
         this->bv.bits_write(begin, end, value);
